// src/interfaces/teamsInterfaces.tsx
// Generated by https://quicktype.io

export interface Team {
  description: string;
  name: string;
  pokemonsSelected: PokemonsSelected[];
  region: Region;
  userId: string;
  key: string;
}

export interface PokemonsSelected {
  id: string;
  name: string;
  picture: string;
}

export interface Region {
  name: string;
  url: string;
}